requirements.txt
Django==2.2.28
esdk-obs-python==3.20.11
PyYAML==5.4
requests==2.31.0
uWSGI==2.0.22
